+++ src/main/java/org/seedstack/jpa/internal/JpaModule.java
@@ -8,6 +8,8 @@
 
 package org.seedstack.jpa.internal;
 
+import static com.google.inject.multibindings.Multibinder.newSetBinder;
+
 import com.google.inject.PrivateModule;
 import com.google.inject.TypeLiteral;
 import com.google.inject.multibindings.Multibinder;
@@ -23,7 +25,6 @@ import org.seedstack.jpa.spi.JpaRepositoryFactory;
 import org.seedstack.seed.core.internal.transaction.TransactionalProxy;
 
 class JpaModule extends PrivateModule {
-
     private final Map<String, EntityManagerFactory> entityManagerFactories;
     private final Map<String, Class<? extends JpaExceptionHandler>> jpaExceptionHandlerClasses;
     private final List<Class<? extends JpaRepositoryFactory>> jpaRepositoryFactories;
@@ -38,51 +39,35 @@ class JpaModule extends PrivateModule {
 
     @Override
     protected void configure() {
-        if (entityManagerFactories != null && !entityManagerFactories.isEmpty()) {
-            EntityManagerLink entityManagerLink = new EntityManagerLink();
+        // EntityManager
+        EntityManagerLink entityManagerLink = new EntityManagerLink();
+        bind(EntityManager.class).toInstance(TransactionalProxy.create(EntityManager.class, entityManagerLink));
+        expose(EntityManager.class);
 
-            bind(EntityManager.class)
-                    .toInstance(TransactionalProxy.create(EntityManager.class, entityManagerLink));
-            expose(EntityManager.class);
+        // Units
+        entityManagerFactories.forEach((key, value) -> bindUnit(key, value, entityManagerLink));
 
-            for (Map.Entry<String, EntityManagerFactory> entry : entityManagerFactories.entrySet()) {
-                bindUnit(entry.getKey(), entry.getValue(), entityManagerLink);
-            }
-
-            Multibinder<JpaRepositoryFactory> jpaRepositoryFactoryMultibinder = Multibinder
-                    .newSetBinder(binder(), JpaRepositoryFactory.class);
-            for (Class<? extends JpaRepositoryFactory> jpaRepositoryFactory : jpaRepositoryFactories) {
-                jpaRepositoryFactoryMultibinder.addBinding().to(jpaRepositoryFactory);
-            }
-
-            expose(new JpaRepositoryFactoriesTypeLiteral());
-        }
+        // JPA repository factories
+        Multibinder<JpaRepositoryFactory> setBinder = newSetBinder(binder(), JpaRepositoryFactory.class);
+        jpaRepositoryFactories.forEach(jpaRepositoryFactory -> setBinder.addBinding().to(jpaRepositoryFactory));
+        expose(new JpaRepositoryFactoriesTypeLiteral());
     }
 
-    private void bindUnit(String name, EntityManagerFactory entityManagerFactory,
-            EntityManagerLink entityManagerLink) {
-        Class<? extends JpaExceptionHandler> unitExceptionHandlerClass = jpaExceptionHandlerClasses
-                .get(name);
+    private void bindUnit(String name, EntityManagerFactory entityManagerFactory, EntityManagerLink entityManagerLink) {
+        Class<? extends JpaExceptionHandler> unitExceptionHandlerClass = jpaExceptionHandlerClasses.get(name);
 
         if (unitExceptionHandlerClass != null) {
-            bind(JpaExceptionHandler.class).annotatedWith(Names.named(name))
-                    .to(unitExceptionHandlerClass);
+            bind(JpaExceptionHandler.class).annotatedWith(Names.named(name)).to(unitExceptionHandlerClass);
         } else {
-            bind(JpaExceptionHandler.class).annotatedWith(Names.named(name))
-                    .toProvider(Providers.of(null));
+            bind(JpaExceptionHandler.class).annotatedWith(Names.named(name)).toProvider(Providers.of(null));
         }
-
-        JpaTransactionHandler transactionHandler = new JpaTransactionHandler(entityManagerLink,
-                entityManagerFactory);
-        bind(JpaTransactionHandler.class).annotatedWith(Names.named(name))
-                .toInstance(transactionHandler);
-
         expose(JpaExceptionHandler.class).annotatedWith(Names.named(name));
+
+        JpaTransactionHandler transactionHandler = new JpaTransactionHandler(entityManagerLink, entityManagerFactory);
+        bind(JpaTransactionHandler.class).annotatedWith(Names.named(name)).toInstance(transactionHandler);
         expose(JpaTransactionHandler.class).annotatedWith(Names.named(name));
     }
 
-    private static class JpaRepositoryFactoriesTypeLiteral extends
-            TypeLiteral<Set<JpaRepositoryFactory>> {
-
+    private static class JpaRepositoryFactoriesTypeLiteral extends TypeLiteral<Set<JpaRepositoryFactory>> {
     }
 }